split the app into two gradle projects
split the client and server parts into two separate gradle projects, makes running and integrating gmail api easier in the long run
