--- build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'org.example'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    compile 'com.google.apis:google-api-services-gmail:v1-rev20211108-1.32.1'
-}
-
-test {
-    useJUnitPlatform()
-}
