+++ server/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sat Apr 01 01:17:31 EDT 2023
+gradle.version=7.4
